Parsers: parser_async -> parser, non-async removed
